third_party/blink/web_tests/svg/dynamic-updates/SVGFESpotLightElement-svgdom-specularExponent-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'specularExponent' property of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS spotLightElement.specularExponent.baseVal is 20
PASS spotLightElement.specularExponent.baseVal is 1
PASS successfullyParsed is true

TEST COMPLETE
